requirements.txt
fastapi
uvicorn
pydantic>=2.0.0
google-genai
langfuse
pandas
yfinance
fredapi
python-dotenv
beautifulsoup4
requests
newsapi-python
